requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.2
pydantic>=2.0
fastapi>=0.90.0
uvicorn>=0.20.0
asyncio

# Dashboard service
streamlit>=1.24.0
pydeck>=0.8.0
plotly>=5.14.0

# Data processing
pandas>=2.0.0
numpy>=1.24.3

# Data scraping
beautifulsoup4>=4.11.2
fake-useragent>=1.1.1
schedule>=1.1.0

# Financial analysis
openai>=1.0.0

# Model auditing
scikit-learn>=1.2.2
matplotlib>=3.7.1
seaborn>=0.12.2
joblib>=1.2.0

# AI Tools Directory
importlib-metadata>=6.0.0

# Web automation
playwright>=1.40.0
